Rename person_search_gender_query_builder_helper to person_search_by_sex_at_birth_query_builder_helper;
