+++ spec/support/helpers/person_search_by_sex_at_birth_query_builder_helper.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+module PersonSearchBySexAtBirthQueryBuilderHelper
+  def gender_query
+    query = {
+      "bool": {
+        "must": [
+          {
+            "query_string": {
+              "default_field": 'gender',
+              "query": 'male',
+              "boost": '1'
+            }
+          }
+        ]
+      }
+    }
+
+    build_query(query).as_json
+  end
+end
