+++ spec/support/person_search_result_builder.rb
@@ -11,7 +11,7 @@ require 'support/helpers/person_search_by_last_first_name_query_builder_part_one
 require 'support/helpers/person_search_by_last_first_name_query_builder_part_two_helper'
 require 'support/helpers/person_search_by_date_of_birth_query_builder_helper'
 require 'support/helpers/person_search_by_approximate_age_query_builder_helper'
-require 'support/helpers/person_search_gender_query_builder_helper'
+require 'support/helpers/person_search_by_sex_at_birth_query_builder_helper'
 
 class PersonSearchResultBuilder
   include QueryBuilderHelper
@@ -24,7 +24,7 @@ class PersonSearchResultBuilder
   include PersonSearchByLastFirstNameQueryBuilderPartTwoHelper
   include PersonSearchByDateOfBirthQueryBuilderHelper
   include PersonSearchByApproximateAgeQueryBuilderHelper
-  include PersonSearchGenderQueryBuilderHelper
+  include PersonSearchBySexAtBirthQueryBuilderHelper
 
   attr_reader :search_result
